+++ app/identifiers/AnyMoreChangesId.scala
@@ -14,16 +14,10 @@
  * limitations under the License.
  */
 
-package forms.vary
+package identifiers
 
-import forms.mappings.Mappings
-import javax.inject.Inject
-import play.api.data.Form
+case object AnyMoreChangesId extends TypedIdentifier[Boolean] {
+  self =>
+  override def toString: String = "anyMoreChanges"
 
-class AnyMoreChangesFormProvider @Inject() extends Mappings {
-
-  def apply(): Form[Boolean] =
-    Form(
-      "value" -> boolean("messages__any_more_changes__error__required")
-    )
 }
