+++ app/views/anyMoreChanges.scala.html
@@ -16,7 +16,7 @@
 
 @import config.FrontendAppConfig
 @import uk.gov.hmrc.play.views.html._
-@import controllers.vary.routes._
+@import controllers.routes._
 @import utils.Constants._
 @import play.api.mvc.Call
