--- conf/vary.routes
@@ -1,5 +1,0 @@
-# microservice specific routes
-
-GET        /make-more-changes/:srn    controllers.vary.AnyMoreChangesController.onPageLoad(srn: Option[String])
-POST       /make-more-changes/:srn    controllers.vary.AnyMoreChangesController.onSubmit(srn: Option[String])
-
