+++ test/navigators/EstablishersIndividualNavigatorSpec.scala
@@ -108,7 +108,7 @@ object EstablishersIndividualNavigatorSpec extends OptionValues {
 
   private def checkYourAnswers(mode: Mode) = controllers.register.establishers.individual.routes.CheckYourAnswersController.onPageLoad(mode, 0, None)
 
-  private def anyMoreChanges = controllers.vary.routes.AnyMoreChangesController.onPageLoad(None)
+  private def anyMoreChanges = controllers.routes.AnyMoreChangesController.onPageLoad(None)
 
   private def exitJourney(mode: Mode) = if (mode == NormalMode) checkYourAnswers(mode) else anyMoreChanges
